Changed price from integer to decimal

+++ src/Application/DTO/ClothingAccountDto.cs
@@ -5,8 +5,8 @@ namespace Application.DTO
     public class ClothingAccountDto : BaseDto
     {
         public string Name { get; set; }
-        public int OldPrice { get; set; }
-        public int ValidPrice { get; set; }
+        public decimal OldPrice { get; set; }
+        public decimal ValidPrice { get; set; }
         public DateTime CreationDate { get; set; }
         public HashSet<ImageDto> Images { get; set; }
         public SizeDto Size { get; set; }
