+++ src/Core/Entities/Clothing.cs
@@ -8,8 +8,8 @@ namespace Core.Entities
         public string Descriprion { get; set; }
         public int SizeId { get; set; }
         public int TypeId { get; set; }
-        public int? ValidPrice { get; set; }
-        public int? OldPrice { get; set; }
+        public decimal? ValidPrice { get; set; }
+        public decimal? OldPrice { get; set; }
         public DateTime CreationDate { get; set; }
         public DateTime? DayOfSold { get; set; }
         public string ApplicationUserId { get; set; }
