+++ src/Web/ViewModels/ClothingAccountViewModel.cs
@@ -5,8 +5,8 @@ namespace Web.ViewModels
     public class ClothingAccountViewModel : BaseViewModel
     {
         public string Name { get; set; }
-        public int OldPrice { get; set; }
-        public int ValidPrice { get; set; }
+        public decimal OldPrice { get; set; }
+        public decimal ValidPrice { get; set; }
         public DateTime CreationDate { get; set; }
         public string SizeName { get; set; }
         public string TypeName { get; set; }
