+++ src/Web/ViewModels/ClothingChangePriceViewModel.cs
@@ -5,10 +5,10 @@ namespace Web.ViewModels
     public class ClothingChangePriceViewModel
     {
         [Display(Name = "OldPrice")]
-        public int? OldPrice { get; set; }
+        public decimal? OldPrice { get; set; }
 
         [Required]
         [Display(Name = "ValidPrice")]
-        public int ValidPrice { get; set; }
+        public decimal ValidPrice { get; set; }
     }
 }
